feat(BarChartUtils): update Variant import path
The changes update the import path for the `Variant` type from `../BarChartV2` to `../types`. This change ensures that the `Variant` type is imported from the correct location, making the code more maintainable and easier to understand.

feat(MiniBarChart): optimize MiniBarChart component

The changes in the `MiniBarChart` component optimize its performance and simplify the code:

- Use `useMemo` hooks to memoize the calculation of `dataKeys`, `colors`, and `chartConfig`, improving performance by avoiding unnecessary re-calculations.
- Remove the `useEffect` hook and `ResizeObserver` implementation, as the component now uses a fixed size instead of dynamically adjusting to the container size.
- Add a new `size` prop to allow users to control the size of the chart.
- Use the `useTheme` hook to determine the appropriate color for the internal bar line based on the current theme (light or dark).